app/serper/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
aiohttp>=3.8.5
beautifulsoup4>=4.12.2
playwright>=1.39.0
transformers>=4.35.0
torch>=2.1.0
numpy>=1.24.3
huggingface-hub>=0.17.3
pydantic>=2.4.2
